--- android/common/bin/src/main/res/layout/root_navigation_host.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-    android:name="androidx.navigation.fragment.NavHostFragment"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    app:defaultNavHost="true"
-    android:id="@+id/rootNavigationHost"
-    app:navGraph="@navigation/root"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-</fragment>
