--- android/common/bin/src/main/res/navigation/root.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-            xmlns:app="http://schemas.android.com/apk/res-auto"
-            android:id="@+id/root" app:startDestination="@id/marketingFragment">
-
-    <fragment android:id="@+id/marketingFragment"
-              android:name="com.hedvig.android.owldroid.ui.marketing.MarketingFragment"
-              android:label="MarketingFragment">
-        <action android:id="@+id/action_marketingFragment_to_chatFragment" app:destination="@id/chatFragment"/>
-    </fragment>
-    <fragment android:id="@+id/chatFragment" android:name="com.hedvig.app.react.chat.ChatFragment"
-              android:label="ChatFragment">
-        <action android:id="@+id/action_chatFragment_to_offerFragment" app:destination="@id/offerFragment"/>
-        <action android:id="@+id/action_chatFragment_to_logged_in_navigation"
-                app:destination="@id/logged_in_navigation"/>
-    </fragment>
-    <fragment android:id="@+id/offerFragment" android:name="com.hedvig.app.react.offer.OfferFragment"
-              android:label="OfferFragment">
-        <action android:id="@+id/action_offerFragment_to_chatFragment" app:destination="@id/chatFragment"/>
-    </fragment>
-    <include app:graph="@navigation/logged_in_navigation"/>
-</navigation>
