+++ android/common/src/main/res/navigation/logged_in_navigation.xml
@@ -120,7 +120,7 @@
 
     <fragment
         android:id="@+id/trustlyFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.payment.TrustlyFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.payment.TrustlyFragment"
         android:label="TrustlyFragment"/>
 
     <fragment
@@ -130,7 +130,7 @@
 
     <fragment
         android:id="@+id/claimsCommonClaimFragment"
-        android:name="com.hedvig.android.owldroid.ui.claims.commonclaim.CommonClaimFragment"
+        android:name="com.hedvig.android.owldroid.feature.claims.ui.commonclaim.CommonClaimFragment"
         android:label="QuickActionFragment">
 
         <action
@@ -146,7 +146,7 @@
 
     <fragment
         android:id="@+id/claimsEmergencyFragment"
-        android:name="com.hedvig.android.owldroid.ui.claims.commonclaim.EmergencyFragment"
+        android:name="com.hedvig.android.owldroid.feature.claims.ui.commonclaim.EmergencyFragment"
         android:label="QuickActionFragment">
 
         <action
@@ -162,17 +162,17 @@
 
     <fragment
         android:id="@+id/referralFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.referral.ReferralFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.referral.ReferralFragment"
         android:label="ReferralFragment"/>
 
     <fragment
         android:id="@+id/myInfoFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.myinfo.MyInfoFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.myinfo.MyInfoFragment"
         android:label="MyInfoFragment"/>
 
     <fragment
         android:id="@+id/myHomeFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.myhome.MyHomeFragment"
+        android:name="com.hedvig.android.owldroid.featuer.profile.ui.myhome.MyHomeFragment"
         android:label="MyHomeFragment">
 
         <action
@@ -188,17 +188,17 @@
 
     <fragment
         android:id="@+id/coinsuredFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.coinsured.CoinsuredFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.coinsured.CoinsuredFragment"
         android:label="CoinsuredFragment"/>
 
     <fragment
         android:id="@+id/charityFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.charity.CharityFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.charity.CharityFragment"
         android:label="CharityFragment"/>
 
     <fragment
         android:id="@+id/paymentFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.payment.PaymentFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.payment.PaymentFragment"
         android:label="PaymentFragment">
 
         <action
@@ -213,12 +213,12 @@
 
     <fragment
         android:id="@+id/feedbackFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.feedback.FeedbackFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.feedback.FeedbackFragment"
         android:label="FeedbackFragment"/>
 
     <fragment
         android:id="@+id/aboutAppFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.aboutapp.AboutAppFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.aboutapp.AboutAppFragment"
         android:label="AboutAppFragment">
 
         <action
@@ -233,6 +233,6 @@
 
     <fragment
         android:id="@+id/licensesFragment"
-        android:name="com.hedvig.android.owldroid.ui.profile.aboutapp.LicensesFragment"
+        android:name="com.hedvig.android.owldroid.feature.profile.ui.aboutapp.LicensesFragment"
         android:label="LicensesFragment"/>
 </navigation>
