+++ android/common/src/main/res/navigation/root.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/marketingFragment"
-        android:name="com.hedvig.android.owldroid.ui.marketing.MarketingFragment"
+        android:name="com.hedvig.android.owldroid.feature.marketing.ui.MarketingFragment"
         android:label="MarketingFragment">
         <action
             android:id="@+id/action_marketingFragment_to_chatFragment"
